Ревизия кода: вынесены функции Google Maps
