+++ AGENTS.md
@@ -258,4 +258,5 @@
 - Общая функция handleValidation заменяет проверки в контроллерах, файл tasks.js удалён
 - В tasks.service.ts и services/tasks.js дублирование расчёта маршрута заменено функцией applyRouteInfo
 - Удалён дублирующийся скрипт .husky/husky.sh
+- Функции Google Maps вынесены в модуль `bot/shared/mapUtils.js`
