+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   заголовка Authorization
 - Страница `/cp/logs` заменяет `/cp/admin` и показывает логи
 - Система логирования переведена на WG Log Engine
+- Общие функции Google Maps собраны в `bot/shared/mapUtils.js`
 - Добавлена цветовая маркировка уровней и уведомления Telegram через движок
 - Запрос `updateTask` фильтрует поля и игнорирует ключи с `$`, устранена ошибка CodeQL
 - Система ролей дополнена числовыми масками доступа; middleware `checkRole` поддерживает проверку масок
