+++ README.md
@@ -10,6 +10,7 @@
 - Веб‑панель администратора на базе TailAdmin.
 - REST API с документацией Swagger по пути `/api-docs`.
 - Поддержка расчёта маршрутов через сервис OSRM.
+- Общие функции Google Maps вынесены в `bot/shared/mapUtils.js`.
 - Примеры конфигурации Prometheus лежат в каталоге `prometheus`.
 - Метрики Prometheus по пути `/metrics`, middleware `metrics.ts` считает общее количество запросов и длительность.
 - Проверка подписи initData веб‑приложения на сервере.
