+++ ROADMAP.md
@@ -114,6 +114,7 @@
 98. Страница `/tasks` позволяет скачать таблицу задач в CSV или JSON
 99. Кнопки действий задачи закреплены внизу внутри формы, цвет «Принять» и «Выполнено» зависит от статуса
 100. Команда `/task_form_app` отправляет ссылку на форму создания задачи
+101. Функции работы с Google Maps вынесены в общий модуль `bot/shared/mapUtils.js`
 101. Все изменения задач фиксируются в коллекции логов
 102. `authFetch` перенаправляет на `/login` при отсутствии токена или статусе 401/403
 103. Срок действия JWT увеличен до семи дней
