+++ bot/web/src/utils/createRouteLink.js
@@ -1,5 +1,6 @@
 // Формирование ссылки маршрута Google Maps из координат
-export default function createRouteLink(start, end, mode='driving'){
-  if(!start||!end) return ''
-  return `https://www.google.com/maps/dir/?api=1&origin=${start.lat},${start.lng}&destination=${end.lat},${end.lng}&travelmode=${mode}`
+import { generateRouteLink } from '../../../shared/mapUtils.js';
+
+export default function createRouteLink(start, end, mode = 'driving') {
+  return generateRouteLink(start, end, mode)
 }
